Minor changes; RectF -> Rect, PointF -> Point, decreased the "backbouncing" of the collisions. Fixed the Graphic resize to also update the hitbox correctly.

+++ Android/BlindLabyrinth/app/src/main/java/fi/arcada/prog/blindlabyrinth/Controller.java
@@ -48,7 +48,7 @@ public class Controller {
         arrowLeft.draw(c);
     }
 
-    public void updateDirection(float x, float y) {
+    public void updateDirection(int x, int y) {
         if(arrowUp.contains(x, y)) {
             direction.y = -1;
         } else if(arrowDown.contains(x, y)) {
